+++ src/gpio/hal_unproven.rs
@@ -18,7 +18,10 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
+use std::time::Duration;
+
 use embedded_hal::digital;
+use embedded_hal::Pwm;
 
 use super::{InputPin, IoPin, Level, OutputPin, Pin};
 
@@ -91,5 +94,121 @@ impl digital::ToggleableOutputPin for IoPin {
 impl digital::ToggleableOutputPin for OutputPin {
     fn toggle(&mut self) {
         OutputPin::toggle(self);
+    }
+}
+
+impl Pwm for OutputPin {
+    type Duty = f64;
+    type Channel = ();
+    type Time = Duration;
+
+    /// Disables a PWM `channel`
+    fn disable(&mut self, _channel: Self::Channel) {
+        let _ = self.clear_pwm();
+    }
+
+    /// Enables a PWM `channel`
+    fn enable(&mut self, _channel: Self::Channel) {
+        let _ = self.set_pwm_frequency(self.frequency, self.duty_cycle);
+    }
+
+    /// Returns the current PWM period
+    fn get_period(&self) -> Self::Time {
+        Duration::from_nanos(if self.frequency == 0.0 {
+            0
+        } else {
+            ((1.0 / self.frequency) * 1_000_000_000.0) as u64
+        })
+    }
+
+    /// Returns the current duty cycle
+    fn get_duty(&self, _channel: Self::Channel) -> Self::Duty {
+        self.duty_cycle
+    }
+
+    /// Returns the maximum duty cycle value
+    fn get_max_duty(&self) -> Self::Duty {
+        1.0
+    }
+
+    /// Sets a new duty cycle
+    fn set_duty(&mut self, _channel: Self::Channel, duty: Self::Duty) {
+        self.duty_cycle = duty.max(0.0).min(1.0);
+
+        if self.soft_pwm.is_some() {
+            let _ = self.set_pwm_frequency(self.frequency, self.duty_cycle);
+        }
+    }
+
+    /// Sets a new PWM period
+    fn set_period<P>(&mut self, period: P)
+    where
+        P: Into<Self::Time>,
+    {
+        let period = period.into();
+        self.frequency =
+            1.0 / (period.as_secs() as f64 + (f64::from(period.subsec_nanos()) / 1_000_000_000.0));
+
+        if self.soft_pwm.is_some() {
+            let _ = self.set_pwm_frequency(self.frequency, self.duty_cycle);
+        }
+    }
+}
+
+impl Pwm for IoPin {
+    type Duty = f64;
+    type Channel = ();
+    type Time = Duration;
+
+    /// Disables a PWM `channel`
+    fn disable(&mut self, _channel: Self::Channel) {
+        let _ = self.clear_pwm();
+    }
+
+    /// Enables a PWM `channel`
+    fn enable(&mut self, _channel: Self::Channel) {
+        let _ = self.set_pwm_frequency(self.frequency, self.duty_cycle);
+    }
+
+    /// Returns the current PWM period
+    fn get_period(&self) -> Self::Time {
+        Duration::from_nanos(if self.frequency == 0.0 {
+            0
+        } else {
+            ((1.0 / self.frequency) * 1_000_000_000.0) as u64
+        })
+    }
+
+    /// Returns the current duty cycle
+    fn get_duty(&self, _channel: Self::Channel) -> Self::Duty {
+        self.duty_cycle
+    }
+
+    /// Returns the maximum duty cycle value
+    fn get_max_duty(&self) -> Self::Duty {
+        1.0
+    }
+
+    /// Sets a new duty cycle
+    fn set_duty(&mut self, _channel: Self::Channel, duty: Self::Duty) {
+        self.duty_cycle = duty.max(0.0).min(1.0);
+
+        if self.soft_pwm.is_some() {
+            let _ = self.set_pwm_frequency(self.frequency, self.duty_cycle);
+        }
+    }
+
+    /// Sets a new PWM period
+    fn set_period<P>(&mut self, period: P)
+    where
+        P: Into<Self::Time>,
+    {
+        let period = period.into();
+        self.frequency =
+            1.0 / (period.as_secs() as f64 + (f64::from(period.subsec_nanos()) / 1_000_000_000.0));
+
+        if self.soft_pwm.is_some() {
+            let _ = self.set_pwm_frequency(self.frequency, self.duty_cycle);
+        }
     }
 }