+++ CHANGELOG.md
@@ -4,8 +4,8 @@
 
 * Add `hal-unproven` feature flag (disabled by default), which includes the relevant `unproven` `embedded-hal` trait implementations. Note that `embedded-hal` may introduce breaking changes for `unproven` traits in patch releases.
 * **Gpio**: Implement `Sync` trait for `IoPin` and `OutputPin`.
-* **Gpio**: Implement `embedded-hal` trait `digital::InputPin` for `Pin`, `InputPin`, `OutputPin` and `IoPin`.
-* **Gpio**: Implement `embedded-hal` traits `digital::{StatefulOutputPin, ToggleableOutputPin}` for `OutputPin` and `IoPin`.
+* **Gpio**: Implement `unproven` `embedded-hal` trait `digital::InputPin` for `Pin`, `InputPin`, `OutputPin` and `IoPin`.
+* **Gpio**: Implement `unproven` `embedded-hal` traits `digital::{StatefulOutputPin, ToggleableOutputPin}` and `Pwm` for `OutputPin` and `IoPin`.
 * **Pwm**: Implement `Display` trait for `Channel` and `Polarity`.
 * **Spi**: Implement `Display` trait for `BitOrder`, `Bus`, `Mode`, `Polarity` and `SlaveSelect`.
 * **Uart**: Add support for the PL011 and mini UART peripherals, USB to serial adapters, XON/XOFF software flow control and RTS/CTS hardware flow control.